fix: rename default css file
