+++ src/styledComponents/StyledFooter.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode } from 'react';
-import styles from '../styles/default.mod.css';
+import styles from '../styles/default.css';
 
 const StyledFooter = ({ children }: { children: ReactNode }) => {
   return <div className={styles['styled-footer']}>{children}</div>;
